\name{RcppExamples-package}
\alias{RcppExamples-package}
\alias{RcppExamples}
\docType{package}
\title{
  Examples for the Rcpp R/C++ Interface library
}
\description{
  This package shows some simple examples for the use of \pkg{Rcpp}.
  Both the older ('classic') and new API are documented.    

  It can also serve as a working template to create packages that use
  \pkg{Rcpp} to interface C++ code or libraries.
}
\details{
  The \pkg{Rcpp} package provides a number of C++ classes that ease
  access to C++ from R. This comprises both passing parameters to
  functions, as well as returning results back from C++ to R.

  Two APIs are supported. The first is an older API which was first
  introduced mostly in 2006 and extended in 2008. This interface is used
  by a few other packages and will be supported going forward.,
  A second and newer API that was started in 2009 offers more
  functionality.

  The \pkg{RcppExamples} package provides some simple examples for use
  of \pkg{Rcpp}. At this point the documentation is not complete in the
  sense of not covering all accessible classes. However, several basic
  use cases are illustrated,
}
\author{Dominick Samperi wrote the initial versions of Rcpp (and
  RcppTemplate) during 2005 and 2006.  Dirk Eddelbuettel made some
  additions, and became maintainer in 2008. Dirk Eddelbuettel and Romain
  Francois have been extending Rcpp since 2009.
}
\keyword{package}

