\name{RcppDataFrame}
\alias{RcppDataFrame}
\title{Rcpp::DataFrame example for Rcpp}
\description{
  A \code{DataFrame} can be passed C++ and can be instantiated as a
  corresponding C++ object using the  Rcpp API. 

  This example shows (in the corresponding C++ code) how to access,
  modify and create a data frame.
}
\details{
  Usage of \code{Rcpp::DataFrame} is fully defined in
  the respective header file. 
}
\author{Dirk Eddelbuettel and Romain Francois}
\examples{
  \dontrun{
  RcppDataFrame()
  }
}
\keyword{programming}
\keyword{interface}
